chore: move scripts and script descriptions dir into scripts directory
