+++ cicd/scripts/milestone-exists.ts
@@ -1,6 +1,6 @@
-import { MilestoneClient } from "./core/MilestoneClient.ts";
-import { ScriptDescriptions } from "./core/ScriptDescriptions.ts";
-import { Utils } from "./core/Utils.ts";
+import { MilestoneClient } from "../core/MilestoneClient.ts";
+import { ScriptDescriptions } from "../core/ScriptDescriptions.ts";
+import { Utils } from "../core/Utils.ts";
 
 const scriptName = Utils.getScriptName()
 const scriptDescriptions: ScriptDescriptions = new ScriptDescriptions();
