+++ cicd/scripts/release-notes-exist.ts
@@ -1,6 +1,6 @@
-import { Utils } from "./core/Utils.ts";
-import { File } from "./core/File.ts";
-import { ScriptDescriptions } from "./core/ScriptDescriptions.ts";
+import { Utils } from "../core/Utils.ts";
+import { File } from "../core/File.ts";
+import { ScriptDescriptions } from "../core/ScriptDescriptions.ts";
 
 const scriptName = Utils.getScriptName()
 const scriptDescriptions: ScriptDescriptions = new ScriptDescriptions();
