+++ cicd/scripts/script-descriptions/release-notes-exist.txt
