+++ cicd/scripts/script-descriptions/validate-release-notes.txt
